requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
mysql-connector-python==8.0.30
Werkzeug==2.2.2
